Apply xrootd-3.1.0-narrowing-conversion.patch

+++ src/XrdFrm/XrdFrmConfig.cc
@@ -1165,7 +1165,7 @@ int XrdFrmConfig::Grab(const char *var, char **Dest, int nosubs)
 XrdOucTList *XrdFrmConfig::InsertPL(XrdOucTList *pL, const char *Path,
                                     int Plen, int isRW)
 {
-   short sval[4] = {isRW, Plen};
+   int sval[4] = {isRW, Plen};
    XrdOucTList *pP = 0, *tP = pL;
 
 // Find insertion point
